make RAdam accept **kwargs in its constructor
This implementation of RAdam dynamically creates a field called 'buffer' in its params. The way fairscale sharded training is implemented in PL is that it gathers all the parameters of the optimizer and passes them to the constructor of a special sharded optimizer wrapper. When this happens, the 'buffer' field is then passed into the constructor of RAdam, but unfortunately it does not accept 'buffer', causing an error. As a quick workaround, just change the constructor to accept kwargs so that it'd just ignore the extra 'buffer' field